move grouping utils into sub folder
