+++ src/utils/grouping/groupCheckersByWeekDay.ts
@@ -1,6 +1,6 @@
 import { isMonday, isTuesday, isWednesday, isThursday, isFriday, isSaturday, isSunday } from 'date-fns';
-import { Checker, CheckerByWeekDay, CheckerFields } from '../interfaces/Checker';
-import { Weekday } from '../interfaces/Weekdays';
+import { Checker, CheckerByWeekDay, CheckerFields } from '../../interfaces/Checker';
+import { Weekday } from '../../interfaces/Weekdays';
 
 export const groupCheckersByWeekDay = (checkers: Checker[]): CheckerByWeekDay[] => {
   let countMonday = 0;
